requirements.txt
click==8.1.7
crewai==0.130.0
fastapi==0.115.9
openai>=1.68.2
pydantic==2.7.4
langchain-core==0.1.52
jsonschema==4.22.0
Jinja2==3.1.4
protobuf==4.25.3
python-dotenv
uvicorn